Training the model
This is the Google Colab code to train the deep learning model using YOLOv3 algorithm. Labelled images of the weapons of 3 classes - handguns, heavy guns and knives are uploaded to drive and the model is trained. It generates test and train data and generates the weight file and the cfg file of YOLOv3