Weapon Detection program
This program is used to detect weapons and label them based on their classes. OpenCV is used to obtain input video from either webcam or internal file. Once the weapon is detected in the frame, it classifies the type of weapon and gives an alert if the weapon is present in the frame.